Develop

+++ include/PathMapHeader.h
@@ -38,7 +38,7 @@ private:
 
     double grid_spacing_;
     double max_path_length_;
-    int neighbor_radius_;
+    double neighbor_radius_;
     double obstacle_threshold_;
 
     IMP::em::DensityHeader density_header_;
@@ -63,7 +63,7 @@ public:
     PathMapHeader(
             double max_path_length = 10.0,
             double grid_spacing = 1.0,
-            int neighbor_radius = 2,
+            double neighbor_radius = 2,
             double obstacle_threshold = std::numeric_limits<double>::epsilon()
     );
 
@@ -77,6 +77,10 @@ public:
      */
     void update_map_dimensions(int nx=-1, int ny=-1, int nz=-1);
 
+    /**
+     * @brief Sets the origin of the path.
+     * @param v The origin vector.
+     */
     void set_path_origin(const IMP::algebra::Vector3D &v);
 
     //! Returns position of the labeling site
@@ -84,25 +88,52 @@ public:
         return path_origin_;
     }
 
-    /// Maximum linker/path length from origin
+    /**
+     * @brief Get the maximum linker/path length from origin.
+     * @return The maximum linker/path length.
+     */
     double get_max_path_length(){
         return max_path_length_;
     }
 
+    /**
+     * @brief Get the simulation grid resolution.
+     * @return The simulation grid resolution as a double.
+     */
     double get_simulation_grid_resolution();
 
+    /**
+     * @brief Set the obstacle threshold.
+     * @param obstacle_threshold The obstacle threshold value
+     */
     void set_obstacle_threshold(double obstacle_threshold);
 
+    /**
+     * @brief Get the obstacle threshold.
+     * @return The obstacle threshold value
+     */
     double get_obstacle_threshold() const{
         return obstacle_threshold_;
     }
 
+    /**
+     * @brief Set the neighbor radius.
+     * @param neighbor_radius The neighbor radius value
+     */
     void set_neighbor_radius(double neighbor_radius);
 
-    int get_neighbor_radius() const{
+    /**
+     * @brief Get the neighbor radius.
+     * @return The neighbor radius as a double.
+     */
+    double get_neighbor_radius() const{
         return neighbor_radius_;
     }
 
+    /**
+     * @brief Get the size of the neighbor box.
+     * @return The size of the neighbor box
+     */
     int get_neighbor_box_size() const;
 
     //! Returns a read-only pointer to the header of the map
@@ -116,11 +147,14 @@ public:
     //! Get origin on the PathMap (the corner of the grid)
     IMP::algebra::Vector3D get_origin() const ;
 
+    /**
+     * @brief Get the edge length of the grid.
+     * @return The edge length of the grid as a double.
+     */
     double get_grid_edge_length();
 
     //! Set origin on the PathMap (the corner of the grid)
     void set_origin(float x, float y, float z);
-
 
 };
